Implement task_SerialParser, and Static String Memory Pool
The SerialParser task passes fully received newline-terminated
messages from the UART (connected to the Jetson SBC) to the
JSONParser task. Its job is to ensure the received serial messages
are appropriately stored in memory until the JSON parsing is complete.

A dedicated Memory Pool is implemented for these strings received over
serial. This avoids constant memory allocations to heap memory to pass
received strings across tasks. The benefit of this approach is to leave
the generic heap memory for other general purpose applications, and to
avoid unnecessary memory fragmentation caused by allocating strings.

The downside is that a large chunk of RAM memory is set aside already
"allocated" for these strings, and that the stored strings are of
limited length (which also affects the static memory footprint). However
static code analysis shows that we have a surplus of RAM even with this
statically-allocated memory, and we can enforce strict rules for
receiving serial messages of max lengths.

Once a message is fully received (the terminating newline is received),
the pointer to the entire string buffer memory pool entry is passed into
the jsonParse queue, which will awaken the JSONParser task.  Passing the
memory pool entry to the JSONParser allows the JSONParser task to free
the char buffer containing the serial message by giving the semaphore
to the pool entry. This marks the string pool entry as unallocated.